Experiment with mypy (failed)
Decided to give mypy a spin to check type annotations. PySide6 types for objects are difficult to get annotated correctly (at least with the way i'm using them). Kept a few of the type annotations that I added  for some basic types in objects for examples etc. Updated some comments for the conversion piece and moved to using sxl's implementation for col2num over the lambda versions I ported.