Handle empty database.
Added code to handle empty database.
Limited 'all' from being inserted as a metadata tag in the database since the application uses all as a special command for tags.
Set minimum length for the metadata combo box.
Commented out some testing code that was left in. Leaving it in comments for future testing.